requirements.txt
appdirs==1.4.4
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cycler==0.10.0
distlib==0.3.1
docopt==0.6.2
filelock==3.0.12
idna==2.10
importlib-metadata==1.7.0
joblib==0.16.0
kiwisolver==1.2.0
matplotlib==3.3.0
nltk==3.5
numpy==1.19.1
objectpath==0.6.1
pandas==1.0.1
Pillow==7.2.0
pipenv==2020.6.2
pipreqs==0.4.10
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
regex==2020.7.14
requests==2.24.0
six==1.15.0
tqdm==4.48.0
urllib3==1.25.9
virtualenv==20.0.27
virtualenv-clone==0.5.4
wordcloud==1.7.0
yarg==0.1.9
zipp==3.1.0
